feat(week2): SharedPreference Add User List
