+++ app/src/main/java/org/sopt/doeuijin/feature/home/HomeFragment.kt
@@ -44,7 +44,7 @@ class HomeFragment : Fragment() {
 
     private fun collectState() {
         activityViewModel.state.flowWithLifecycle(viewLifeCycle).onEach {
-            (binding.rvHome.adapter as? ProfileAdapter)?.submitList(it.profileList)
+            (binding.rvHome.adapter as? ProfileAdapter)?.submitList(it.userList)
         }.launchIn(viewLifeCycleScope)
     }
