+++ app/src/main/java/org/sopt/doeuijin/feature/home/profile/Profile.kt
@@ -1,5 +1,7 @@
 package org.sopt.doeuijin.feature.home.profile
 
+import kotlinx.serialization.Serializable
+
 sealed class Profile {
     abstract val name: String
     abstract val description: String
@@ -11,6 +13,7 @@ sealed class Profile {
         override val image: String,
     ) : Profile()
 
+    @Serializable
     data class FriendProfile(
         override val name: String,
         override val description: String,
