+++ core/common/src/main/java/org/sopt/common/key/SharedPreferenceKey.kt
@@ -5,5 +5,7 @@ object SharedPreferenceKey {
     const val USER_PW = "USER_PW"
     const val USER_NAME = "USER_NAME"
 
+    const val USER_LIST = "USER_LIST"
+
     const val AUTO_LOGIN = "AUTO_LOGIN"
 }
